+++ backend/src/baserow/contrib/database/table/models.py
@@ -80,10 +80,13 @@ class TableModelQuerySet(models.QuerySet):
             field_name = field_object["name"]
             model_field = self.model._meta.get_field(field_name)
 
-            sub_filter = field_object["type"].contains_query(
-                field_name, search, model_field, field_object["field"]
-            )
-            filter_builder.filter(sub_filter)
+            try:
+                sub_filter = field_object["type"].contains_query(
+                    field_name, search, model_field, field_object["field"]
+                )
+                filter_builder.filter(sub_filter)
+            except Exception:  # nosec B112
+                continue
 
         return filter_builder.apply_to_queryset(self)